[feat][alloc] Add parallel H2 allocation
Remove JVM locks for the allocator and make the allocator parallel.
Each thread accesses the allocator and locks the id-to-region
mapping. Thus, threads access different mappings independently.

If there is no mapping, or the mapping does not fit the object,
a global lock protects the region-array so that a thread can
safely get a new region.

Finally, global state is updated through a global lock.