[wip][dbg] Debugging guard for Lost Region Bug
This commit disables the buggy code and all the debugging code so that
jvm runs normally. Parts of this commit will be removed after the
bug is fixed.

Current buggy scenario: during CM somehow a region from h2 is not
marked as live even though it is still reachable. It ends up
reclaiming the region resulting a false execution.